requirements.txt
androidstorage4kivy==0.1.1
appdirs==1.4.4
build==1.2.1
buildozer==1.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Cython==0.29.33
distlib==0.3.8
docutils==0.21.2
ffpyplayer==4.5.1
filelock==3.15.4
idna==3.7
Jinja2==3.1.4
Kivy==2.3.0
Kivy-examples==2.3.0
Kivy-Garden==0.1.5
MarkupSafe==2.1.5
packaging==24.1
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
plyer==2.1.0
ptyprocess==0.7.0
Pygments==2.18.0
pyproject_hooks==1.1.0
python-for-android==2024.1.21
requests==2.31.0
setuptools==72.2.0
sh==1.14.3
toml==0.10.2
urllib3==2.2.1
virtualenv==20.26.3
